backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pynacl==1.5.0
base58==2.1.1
